add an option to scan plugin libraries
it is possible to include Jars inside plugins that may also depend on
Jenkins, so add an option to include those in the scan as well.

This defaults to off to retain compatability with the old version,
however this may not be the correct way if people are looking at the
report of this as gospel.